rename sections to tiles and remove from packages
