+++ __docs__/components/gallery/tiles/modal-tile.tsx
@@ -4,34 +4,27 @@ import {StyleSheet} from "aphrodite";
 import Button from "@khanacademy/wonder-blocks-button";
 import {View} from "@khanacademy/wonder-blocks-core";
 import {OnePaneDialog} from "@khanacademy/wonder-blocks-modal";
-import {Body, HeadingLarge} from "@khanacademy/wonder-blocks-typography";
+import {Body} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../styles";
 
-export default function ModalSection() {
+export default function ModalTile() {
     return (
-        <>
-            <HeadingLarge id="modal" tag="h3" style={styles.sectionLabel}>
-                Modal
-            </HeadingLarge>
-
-            <View style={styles.section}>
-                <ComponentTile
-                    name="OnePaneDialog"
-                    href="/?path=/docs/packages-modal-onepanedialog--docs"
-                    description={`Standard layout for most straightforward
+        <ComponentTile
+            name="OnePaneDialog"
+            href="/?path=/docs/packages-modal-onepanedialog--docs"
+            description={`Standard layout for most straightforward
                         modal experiences. Modals display in front of the
                         current view and prevent interaction background until
                         they are dismissed.`}
-                >
-                    <View style={localStyles.previewSizer}>
-                        <View style={localStyles.modalPositioner}>
-                            <OnePaneDialog
-                                title="Dialog title"
-                                content={
-                                    <Body>
-                                        {`Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+        >
+            <View style={localStyles.previewSizer}>
+                <View style={localStyles.modalPositioner}>
+                    <OnePaneDialog
+                        title="Dialog title"
+                        content={
+                            <Body>
+                                {`Lorem ipsum dolor sit amet, consectetur adipiscing elit,
                 sed do eiusmod tempor incididunt ut labore et dolore magna
                 aliqua. Ut enim ad minim veniam, quis nostrud exercitation
                 ullamco laboris nisi ut aliquip ex ea commodo consequat.
@@ -39,19 +32,17 @@ export default function ModalSection() {
                 esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
                 occaecat cupidatat non proident, sunt in culpa qui officia
                 deserunt mollit anim id est.`}
-                                    </Body>
-                                }
-                                footer={
-                                    <View style={localStyles.footer}>
-                                        <Button>Next</Button>
-                                    </View>
-                                }
-                            />
-                        </View>
-                    </View>
-                </ComponentTile>
+                            </Body>
+                        }
+                        footer={
+                            <View style={localStyles.footer}>
+                                <Button>Next</Button>
+                            </View>
+                        }
+                    />
+                </View>
             </View>
-        </>
+        </ComponentTile>
     );
 }
