+++ __docs__/components/gallery/tiles/switch-tile.tsx
@@ -2,19 +2,14 @@ import * as React from "react";
 
 import {RenderStateRoot, View} from "@khanacademy/wonder-blocks-core";
 import Switch from "@khanacademy/wonder-blocks-switch";
-import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
 import {styles} from "../styles";
 
-export default function BannerSection() {
+export default function SwitchTile() {
     const [checked, setChecked] = React.useState<boolean>(false);
     return (
         <RenderStateRoot>
-            <HeadingLarge id="switch" tag="h3" style={styles.sectionLabel}>
-                Switch
-            </HeadingLarge>
-
             <ComponentTile
                 name="Switch"
                 href="/?path=/docs/packages-switch--docs"
